refactor(codegen): split string concat out of lower_string_method.rs (#7615)
`lower_string_method.rs` was 1,957 lines against the 2,000-line cap, and
the Layer 1 rooting migration adds closure scopes to five of its
functions — `with_operands_rooted` and `with_rooted_accumulator` both
re-indent the body they own, which is line growth on a file with 43 lines
of headroom.

Pure move at the boundary the file already had: everything above
`lower_string_self_append` dispatches a `str.<method>(...)` call,
everything from it down lowers `a + b` / `s += x` on strings.
`str_operand_handle_tag_dispatched` becomes `pub(crate)` because three
dispatch arms above still call it.

lower_string_method.rs 1,957 -> 1,368; lower_string_concat.rs 612.

Also lands the first four module migrations of slice 8 (they share the
`expr/binary.rs` import line with the move):

* `expr/binary.rs` — five `lower_operand_pair_rooted` + `temp_root_release`
  pairs collapse into one `lower_rooted_dynamic_binary` helper over
  `with_operands_rooted`.
* `expr/math_simple.rs` — `MapSet` becomes a `RootedGroup` (two operands,
  unequal windows, eight arm-specific re-read points); `MapGet`/`MapHas`
  become `with_operands_rooted`. `Expr::ArrayMap` gains the root it never
  had: the receiver was lowered, the callback was lowered, and only THEN
  was the receiver unboxed — the unbox sat below its own window.
* `expr/static_field_meta.rs` — `ClassExprFresh` becomes a `RootedGroup`
  over the class object plus a nested `with_rooted_accumulator` for the
  `__perry_ctor_caps` snapshot array, which was threaded through a bare
  SSA register.
* `expr/dyn_extern_i18n.rs` — the namespace-object build becomes
  `with_rooted_accumulator`.
* `lower_call/new.rs` — `refresh_rooted_args` and the
  `temp_root_scope_begin`/`_end` marker become one escaping `RootedGroup`;
  the null marker slot is gone with them.

`RootedGroup::adopt_emitted` gains a `protect` flag (the WINDOW, not the
strategy) and `RootedGroup::is_rooted` returns whether a slot exists.
