+++ crates/perry-codegen/src/lib.rs
@@ -24,6 +24,7 @@ pub(crate) mod loop_purity;
 pub(crate) mod lower_array_method;
 pub(crate) mod lower_call;
 pub(crate) mod lower_conditional;
+pub(crate) mod lower_string_concat;
 pub(crate) mod lower_string_method;
 pub mod module;
 pub mod nanbox;
