+++ crates/perry-codegen/src/lower_string_concat.rs
@@ -1,0 +1,616 @@
+//! String CONCATENATION and self-append lowering.
+//!
+//! Split out of `lower_string_method.rs` (#7615 slice 8). That file was 1,957
+//! lines against `scripts/check_file_size.sh`'s 2,000-line cap, and the Layer 1
+//! rooting migration has to add closure scopes to five of the functions here —
+//! `with_operands_rooted` and `with_rooted_accumulator` both re-indent the body
+//! they own, which is line growth on a file with 43 lines of headroom.
+//!
+//! Pure move, no behaviour change. The boundary is the one the file already
+//! had: everything above it dispatches a `str.<method>(...)` call, everything
+//! here lowers `a + b` / `s += x` on strings. `str_operand_handle_tag_dispatched`
+//! is `pub(crate)` rather than private because three of the dispatch arms above
+//! call it.
+
+use anyhow::{anyhow, Result};
+use perry_hir::Expr;
+
+use crate::expr::{lower_expr, nanbox_string_inline, unbox_str_handle, FnCtx};
+use crate::type_analysis::is_string_expr;
+use crate::types::{DOUBLE, I1, I32, I64};
+
+use crate::expr::temp_root::{
+    lower_exprs_rooted, lower_operand_pair_rooted, temp_root_get_double, temp_root_get_i64,
+    temp_root_push_double, temp_root_push_i64, temp_root_release, temp_root_truncate,
+};
+
+/// Lower the `str = str + rhs` self-append pattern. Uses the in-place
+/// `js_string_append` runtime function (refcount=1 → mutate in place,
+/// otherwise allocate). The returned pointer is stored back to the local
+/// slot — `js_string_append` may realloc when growing past capacity.
+///
+/// This is the load-bearing optimization for the canonical `let str = "";
+/// for (...) str = str + "a"` string-build pattern.
+pub(crate) fn lower_string_self_append(
+    ctx: &mut FnCtx<'_>,
+    local_id: u32,
+    rhs: &Expr,
+) -> Result<String> {
+    let slot = ctx
+        .locals
+        .get(&local_id)
+        .ok_or_else(|| anyhow!("string self-append: local {} not in scope", local_id))?
+        .clone();
+
+    // Representation-selection Phase 3a: canonical-Str destination —
+    // tag-dispatch on the slot bits inline instead of paying the two opaque
+    // `js_get_string_pointer_unified` calls per iteration.
+    if crate::expr::local_is_canonical_str(ctx, local_id) {
+        return lower_canonical_str_self_append(ctx, local_id, rhs, &slot);
+    }
+
+    // Lower the RHS first (might be a string literal, a local, or a
+    // computed expression). For non-string RHS we'd need to coerce, but
+    // the bench_string_ops case always uses a string literal, so for the
+    // first slice we require the RHS to be a known string.
+    if !is_string_expr(ctx, rhs) {
+        // Fall back to the slower concat path: load the local, do a
+        // generic concat-coerce, store back.
+        let lhs_val = ctx.block().load(DOUBLE, &slot);
+        let _lhs = lhs_val.clone();
+        let rhs_val = lower_expr(ctx, rhs)?;
+        let blk = ctx.block();
+        // Issue #214: SSO-safe unbox.
+        let l_handle = unbox_str_handle(blk, &lhs_val);
+        // Coerce non-string RHS to a string handle.
+        let r_handle = blk.call(I64, "js_jsvalue_to_string", &[(DOUBLE, &rhs_val)]);
+        let result = blk.call(
+            I64,
+            "js_string_append",
+            &[(I64, &l_handle), (I64, &r_handle)],
+        );
+        let new_box = nanbox_string_inline(blk, &result);
+        blk.store(DOUBLE, &new_box, &slot);
+        return Ok(new_box);
+    }
+
+    let rhs_box = lower_expr(ctx, rhs)?;
+    let blk = ctx.block();
+    let lhs_box = blk.load(DOUBLE, &slot);
+    // Issue #214: SSO-safe unbox.
+    let l_handle = unbox_str_handle(blk, &lhs_box);
+    let r_handle = unbox_str_handle(blk, &rhs_box);
+    let new_handle = blk.call(
+        I64,
+        "js_string_append",
+        &[(I64, &l_handle), (I64, &r_handle)],
+    );
+    let new_box = nanbox_string_inline(blk, &new_handle);
+    blk.store(DOUBLE, &new_box, &slot);
+    Ok(new_box)
+}
+
+/// Repsel Phase 3a: is this expression PROVEN to lower to a heap-tagged
+/// (`STRING_TAG`) NaN-box — never SSO bits, never a non-string? String
+/// literals load the interned pool handle (`@.str.N.handle`, always a heap
+/// `StringHeader` from `js_string_from_bytes`); `String(x)` routes through
+/// `js_string_coerce`, which always allocates a heap header. Deliberately
+/// NOT included: `Binary Add` string results — the pairwise concat lowering
+/// returns `js_string_concat_box`, which assembles ≤5-byte ASCII results as
+/// SSO bits.
+fn proven_heap_string_operand(_ctx: &FnCtx<'_>, e: &Expr) -> bool {
+    match e {
+        Expr::String(_) | Expr::WtfString(_) | Expr::StringCoerce(_) => true,
+        Expr::Conditional {
+            then_expr,
+            else_expr,
+            ..
+        } => {
+            proven_heap_string_operand(_ctx, then_expr)
+                && proven_heap_string_operand(_ctx, else_expr)
+        }
+        _ => false,
+    }
+}
+
+/// Repsel Phase 3a: operand → raw `StringHeader*` handle for the string
+/// helpers, tag-dispatched:
+///
+/// - proven heap-tagged operand (see `proven_heap_string_operand`) → inline
+///   `bitcast; and POINTER_MASK` — zero calls;
+/// - canonical-Str `LocalGet` → 2-arm dispatch: heap `STRING_TAG` bits →
+///   bare `and POINTER_MASK` (hot arm, no call); anything else (SSO bits,
+///   annotation lie) → the legacy `js_get_string_pointer_unified` (which
+///   materializes SSO — cold);
+/// - everything else (or flag off) → the legacy unified call, unchanged.
+///
+/// #7128: the two arms are on separate knobs, because only the second one is
+/// about a selected representation. The proven-heap arm keys on the operand's
+/// static type and fires with zero canonical-`Str` locals in the program.
+pub(crate) fn str_operand_handle_tag_dispatched(
+    ctx: &mut FnCtx<'_>,
+    object: &Expr,
+    recv_box: &str,
+) -> String {
+    use crate::nanbox::POINTER_MASK_I64;
+    if crate::expr::static_string_lowering_enabled() && proven_heap_string_operand(ctx, object) {
+        let bits = ctx.block().bitcast_double_to_i64(recv_box);
+        return ctx.block().and(I64, &bits, POINTER_MASK_I64);
+    }
+    let canonical = crate::expr::canonical_str_locals_enabled()
+        && matches!(
+            object, Expr::LocalGet(id) if crate::expr::local_is_canonical_str(ctx, *id)
+        );
+    if !canonical {
+        return unbox_str_handle(ctx.block(), recv_box);
+    }
+    let bits = ctx.block().bitcast_double_to_i64(recv_box);
+    let tag = ctx.block().lshr(I64, &bits, "48");
+    let is_heap = ctx
+        .block()
+        .icmp_eq(I64, &tag, crate::nanbox::STRING_TAG_TOP16_I64);
+
+    let heap_idx = ctx.new_block("strrecv.heap");
+    let cold_idx = ctx.new_block("strrecv.cold");
+    let merge_idx = ctx.new_block("strrecv.merge");
+    let heap_label = ctx.block_label(heap_idx);
+    let cold_label = ctx.block_label(cold_idx);
+    let merge_label = ctx.block_label(merge_idx);
+    ctx.block().cond_br(&is_heap, &heap_label, &cold_label);
+
+    ctx.current_block = heap_idx;
+    let h_heap = ctx.block().and(I64, &bits, POINTER_MASK_I64);
+    let heap_pred = ctx.block().label.clone();
+    ctx.block().br(&merge_label);
+
+    ctx.current_block = cold_idx;
+    let h_cold = unbox_str_handle(ctx.block(), recv_box);
+    let cold_pred = ctx.block().label.clone();
+    ctx.block().br(&merge_label);
+
+    ctx.current_block = merge_idx;
+    ctx.block()
+        .phi(I64, &[(&h_heap, &heap_pred), (&h_cold, &cold_pred)])
+}
+
+/// Representation-selection Phase 3a: `s += rhs` for a canonical-Str
+/// destination (`SlotRep::Str` — the `ctx.locals` slot provably holds
+/// NaN-box string bits at rest). Replaces the two opaque
+/// `js_get_string_pointer_unified` calls per iteration with an inline tag
+/// dispatch on the slot bits:
+///
+/// - **both heap** (`STRING_TAG` on both sides): `and POINTER_MASK` →
+///   `js_string_append(h, h)` → `or STRING_TAG` — the hot accumulator-loop
+///   arm; keeps the refcount==1 in-place append (every alias demote site is
+///   untouched by this phase, so `let b = a` still demotes first).
+/// - **both strings, SSO involved**: `js_string_concat_box(box, box)` —
+///   SSO-aware pairwise concat, assembles ≤5-byte ASCII results inline and
+///   never mutates in place. No per-op heap materialization of SSO bits
+///   (RFC §4 "short-string values stay by-value").
+/// - **anything else** (a lying `string` annotation): the exact pre-phase
+///   sequence — `js_get_string_pointer_unified` ×2 (SSO materialize +
+///   number coercion included) → `js_string_append` — so acceptance
+///   behavior is bit-identical to today's on non-string bits (RFC §5.5:
+///   mismatches route to the legacy path, never a new coercion).
+fn lower_canonical_str_self_append(
+    ctx: &mut FnCtx<'_>,
+    _local_id: u32,
+    rhs: &Expr,
+    slot: &str,
+) -> Result<String> {
+    use crate::nanbox::{
+        POINTER_MASK_I64, SHORT_STRING_TAG_TOP16_I64 as TAG_SSO_STR,
+        STRING_TAG_TOP16_I64 as TAG_HEAP_STR,
+    };
+
+    if !is_string_expr(ctx, rhs) {
+        // Non-string rhs: mirror the legacy fallback's evaluation order
+        // (lhs slot load, then rhs), coerce the rhs once (heap handle
+        // guaranteed), then 2-arm on the destination tag only.
+        let lhs_box = ctx.block().load(DOUBLE, slot);
+        // #6951: the load must happen before `rhs` per `s += rhs` evaluation
+        // order (a `rhs` that reassigns `s` must not be observed here), so the
+        // pre-rhs value has to be carried across `rhs`'s evaluation and the
+        // `js_jsvalue_to_string` coercion — both of which allocate. Re-reading
+        // the slot would take the wrong value; re-read the temp root instead.
+        let lhs_root = temp_root_push_double(ctx, &lhs_box);
+        let rhs_val = lower_expr(ctx, rhs)?;
+        let r_handle = ctx
+            .block()
+            .call(I64, "js_jsvalue_to_string", &[(DOUBLE, &rhs_val)]);
+        // The coerced rhs is a bare string handle that has to survive the cold
+        // arm's `unbox_str_handle`, which materializes an SSO destination onto
+        // the heap — another allocation. Root it too and re-read it per arm.
+        let r_root = temp_root_push_i64(ctx, &r_handle);
+        let lhs_box = temp_root_get_double(ctx, &lhs_root);
+        let bits_d = ctx.block().bitcast_double_to_i64(&lhs_box);
+        let tag_d = ctx.block().lshr(I64, &bits_d, "48");
+        let is_heap = ctx.block().icmp_eq(I64, &tag_d, TAG_HEAP_STR);
+
+        let heap_idx = ctx.new_block("strapp.heap");
+        let cold_idx = ctx.new_block("strapp.cold");
+        let merge_idx = ctx.new_block("strapp.merge");
+        let heap_label = ctx.block_label(heap_idx);
+        let cold_label = ctx.block_label(cold_idx);
+        let merge_label = ctx.block_label(merge_idx);
+        ctx.block().cond_br(&is_heap, &heap_label, &cold_label);
+
+        ctx.current_block = heap_idx;
+        let h_d = ctx.block().and(I64, &bits_d, POINTER_MASK_I64);
+        let r_heap = temp_root_get_i64(ctx, &r_root);
+        let h_heap = ctx
+            .block()
+            .call(I64, "js_string_append", &[(I64, &h_d), (I64, &r_heap)]);
+        let heap_pred = ctx.block().label.clone();
+        ctx.block().br(&merge_label);
+
+        ctx.current_block = cold_idx;
+        let h_d2 = unbox_str_handle(ctx.block(), &lhs_box);
+        let r_cold = temp_root_get_i64(ctx, &r_root);
+        let h_cold = ctx
+            .block()
+            .call(I64, "js_string_append", &[(I64, &h_d2), (I64, &r_cold)]);
+        let cold_pred = ctx.block().label.clone();
+        ctx.block().br(&merge_label);
+
+        ctx.current_block = merge_idx;
+        let handle = ctx
+            .block()
+            .phi(I64, &[(&h_heap, &heap_pred), (&h_cold, &cold_pred)]);
+        let new_box = nanbox_string_inline(ctx.block(), &handle);
+        ctx.block().store(DOUBLE, &new_box, slot);
+        // `lhs_root` is the base of the pair, so one truncate drops both. The
+        // index register is defined in the entry block and dominates the merge.
+        temp_root_truncate(ctx, &lhs_root);
+        return Ok(new_box);
+    }
+
+    // Proven-string rhs: mirror the legacy fast path's evaluation order
+    // (rhs first, then the lhs slot load).
+    //
+    // Arm layout — the load-bearing property is that a HEAP destination
+    // ALWAYS reaches `js_string_append` (whose refcount==1 in-place path is
+    // what makes accumulator loops amortized O(n)). Routing a heap-dest /
+    // SSO-rhs iteration through `js_string_concat_box` instead would copy
+    // the whole accumulator every time a ≤5-byte part arrives — O(n²).
+    //
+    //   dest heap, rhs heap  → append(h, h)                 (hot, no calls)
+    //   dest heap, rhs other → append(h, unified(rhs))      (legacy-exact:
+    //                          unified materializes SSO / coerces a lie)
+    //   dest SSO             → js_string_concat_box          (SSO-aware,
+    //                          nothing to mutate in place; result may stay
+    //                          SSO — no per-op heap materialization)
+    //   dest other (lie)     → unified ×2 + append           (legacy-exact)
+    let rhs_box = lower_expr(ctx, rhs)?;
+    let lhs_box = ctx.block().load(DOUBLE, slot);
+    let bits_d = ctx.block().bitcast_double_to_i64(&lhs_box);
+    let bits_r = ctx.block().bitcast_double_to_i64(&rhs_box);
+    let tag_d = ctx.block().lshr(I64, &bits_d, "48");
+    let tag_r = ctx.block().lshr(I64, &bits_r, "48");
+    let d_heap = ctx.block().icmp_eq(I64, &tag_d, TAG_HEAP_STR);
+
+    let dheap_idx = ctx.new_block("strapp.dheap");
+    let heap_idx = ctx.new_block("strapp.heap");
+    let rcold_idx = ctx.new_block("strapp.rcold");
+    let dother_idx = ctx.new_block("strapp.dother");
+    let sso_idx = ctx.new_block("strapp.sso");
+    let cold_idx = ctx.new_block("strapp.cold");
+    let merge_idx = ctx.new_block("strapp.merge");
+    let dheap_label = ctx.block_label(dheap_idx);
+    let heap_label = ctx.block_label(heap_idx);
+    let rcold_label = ctx.block_label(rcold_idx);
+    let dother_label = ctx.block_label(dother_idx);
+    let sso_label = ctx.block_label(sso_idx);
+    let cold_label = ctx.block_label(cold_idx);
+    let merge_label = ctx.block_label(merge_idx);
+    ctx.block().cond_br(&d_heap, &dheap_label, &dother_label);
+
+    // dest heap: split on the rhs tag.
+    ctx.current_block = dheap_idx;
+    let r_heap = ctx.block().icmp_eq(I64, &tag_r, TAG_HEAP_STR);
+    ctx.block().cond_br(&r_heap, &heap_label, &rcold_label);
+
+    ctx.current_block = heap_idx;
+    let h_d = ctx.block().and(I64, &bits_d, POINTER_MASK_I64);
+    let h_r = ctx.block().and(I64, &bits_r, POINTER_MASK_I64);
+    let h_new = ctx
+        .block()
+        .call(I64, "js_string_append", &[(I64, &h_d), (I64, &h_r)]);
+    let box_heap = nanbox_string_inline(ctx.block(), &h_new);
+    let heap_pred = ctx.block().label.clone();
+    ctx.block().br(&merge_label);
+
+    ctx.current_block = rcold_idx;
+    let h_d1 = ctx.block().and(I64, &bits_d, POINTER_MASK_I64);
+    let r_h1 = unbox_str_handle(ctx.block(), &rhs_box);
+    let h_rc = ctx
+        .block()
+        .call(I64, "js_string_append", &[(I64, &h_d1), (I64, &r_h1)]);
+    let box_rcold = nanbox_string_inline(ctx.block(), &h_rc);
+    let rcold_pred = ctx.block().label.clone();
+    ctx.block().br(&merge_label);
+
+    // dest not heap: an SSO dest with a real-string rhs takes the SSO-aware
+    // pairwise concat; a lie on EITHER side keeps the exact legacy sequence
+    // (`js_string_concat_box` treats a non-string operand as empty, but the
+    // legacy unified path ToString-coerces it — `"ab" += 42` must stay
+    // `"ab42"`).
+    ctx.current_block = dother_idx;
+    let d_sso = ctx.block().icmp_eq(I64, &tag_d, TAG_SSO_STR);
+    let r_heap2 = ctx.block().icmp_eq(I64, &tag_r, TAG_HEAP_STR);
+    let r_sso = ctx.block().icmp_eq(I64, &tag_r, TAG_SSO_STR);
+    let r_str = ctx.block().or(I1, &r_heap2, &r_sso);
+    let take_sso = ctx.block().and(I1, &d_sso, &r_str);
+    ctx.block().cond_br(&take_sso, &sso_label, &cold_label);
+
+    ctx.current_block = sso_idx;
+    let box_sso = ctx.block().call(
+        DOUBLE,
+        "js_string_concat_box",
+        &[(DOUBLE, &lhs_box), (DOUBLE, &rhs_box)],
+    );
+    let sso_pred = ctx.block().label.clone();
+    ctx.block().br(&merge_label);
+
+    ctx.current_block = cold_idx;
+    let l_h = unbox_str_handle(ctx.block(), &lhs_box);
+    let r_h = unbox_str_handle(ctx.block(), &rhs_box);
+    let h_cold = ctx
+        .block()
+        .call(I64, "js_string_append", &[(I64, &l_h), (I64, &r_h)]);
+    let box_cold = nanbox_string_inline(ctx.block(), &h_cold);
+    let cold_pred = ctx.block().label.clone();
+    ctx.block().br(&merge_label);
+
+    ctx.current_block = merge_idx;
+    let new_box = ctx.block().phi(
+        DOUBLE,
+        &[
+            (&box_heap, &heap_pred),
+            (&box_rcold, &rcold_pred),
+            (&box_sso, &sso_pred),
+            (&box_cold, &cold_pred),
+        ],
+    );
+    ctx.block().store(DOUBLE, &new_box, slot);
+    Ok(new_box)
+}
+
+/// Lower `string + non_string` (or vice versa) concat with runtime
+/// coercion of the non-string side. The non-string operand passes through
+/// `js_jsvalue_to_string` which inspects its NaN tag and produces the
+/// canonical JS string form (numbers via the formatter at
+/// `crates/perry-runtime/src/value.rs:825`, booleans → `"true"`/`"false"`,
+/// objects → `"[object Object]"`, etc.).
+///
+/// The string-typed side still uses the fast inline `bitcast double → i64;
+/// and POINTER_MASK_I64` unbox; only the non-string side pays the function
+/// call. Both operand handles then feed `js_string_concat`.
+pub(crate) fn lower_string_coerce_concat(
+    ctx: &mut FnCtx<'_>,
+    left: &Expr,
+    right: &Expr,
+    l_is_string: bool,
+    r_is_string: bool,
+) -> Result<String> {
+    // #6951: `l_box` is a heap string in an SSA register while `right` is
+    // lowered. If `right` allocates (`"tag" + f()`), a collection sweeps the
+    // left operand and the concat reads freed memory — a segfault, not a
+    // dropped character. `lower_operand_pair_rooted` emits nothing at all when
+    // `right` provably cannot collect, which is the common `"user_" + i` case.
+    let (l_box, r_box, guard) = lower_operand_pair_rooted(ctx, left, right)?;
+
+    // Issue #58: fused string+value concat — when one side is a string
+    // and the other is not, use the fused runtime call that collapses
+    // js_jsvalue_to_string + js_string_concat into a single allocation
+    // for number operands (the common `"item_" + i` pattern).
+    if l_is_string && !r_is_string {
+        // Issue #214: SSO-safe unbox; repsel Phase 3a: inline `bitcast+and`
+        // for proven-heap operands (string literals — the `"user_" + i`
+        // shape) and tag-dispatch for canonical-Str locals.
+        let l_handle = str_operand_handle_tag_dispatched(ctx, left, &l_box);
+        let blk = ctx.block();
+        let result_handle = blk.call(
+            I64,
+            "js_string_concat_value",
+            &[(I64, &l_handle), (DOUBLE, &r_box)],
+        );
+        let boxed = nanbox_string_inline(blk, &result_handle);
+        temp_root_release(ctx, guard);
+        return Ok(boxed);
+    }
+
+    if !l_is_string && r_is_string {
+        // Issue #214: SSO-safe unbox; repsel Phase 3a: see above.
+        let r_handle = str_operand_handle_tag_dispatched(ctx, right, &r_box);
+        let blk = ctx.block();
+        let result_handle = blk.call(
+            I64,
+            "js_value_concat_string",
+            &[(DOUBLE, &l_box), (I64, &r_handle)],
+        );
+        let boxed = nanbox_string_inline(blk, &result_handle);
+        temp_root_release(ctx, guard);
+        return Ok(boxed);
+    }
+
+    // Both non-string (shouldn't normally reach here) — fall back to
+    // the generic path.
+    let l_handle = ctx
+        .block()
+        .call(I64, "js_jsvalue_to_string", &[(DOUBLE, &l_box)]);
+    // The coercion of the right operand allocates, and `l_handle` is a bare
+    // string address in an SSA register — root it across that call (#6951).
+    let l_root = temp_root_push_i64(ctx, &l_handle);
+    let r_handle = ctx
+        .block()
+        .call(I64, "js_jsvalue_to_string", &[(DOUBLE, &r_box)]);
+    let l_handle = temp_root_get_i64(ctx, &l_root);
+    let blk = ctx.block();
+
+    let result_handle = blk.call(
+        I64,
+        "js_string_concat",
+        &[(I64, &l_handle), (I64, &r_handle)],
+    );
+    let boxed = nanbox_string_inline(blk, &result_handle);
+    temp_root_truncate(ctx, &l_root);
+    temp_root_release(ctx, guard);
+    Ok(boxed)
+}
+
+/// Lower a static `s1 + s2` string concatenation. Both operands must
+/// already be statically string-typed (caller's responsibility — see
+/// `is_string_expr`).
+///
+/// Pattern:
+/// ```llvm
+/// ; %l_box and %r_box are NaN-boxed strings (double values with STRING_TAG)
+/// %l_bits = bitcast double %l_box to i64
+/// %l_handle = and i64 %l_bits, 281474976710655   ; POINTER_MASK_I64
+/// %r_bits = bitcast double %r_box to i64
+/// %r_handle = and i64 %r_bits, 281474976710655
+/// %result_handle = call i64 @js_string_concat(i64 %l_handle, i64 %r_handle)
+/// %result_box = call double @js_nanbox_string(i64 %result_handle)
+/// ```
+///
+/// The bitcast+and is the inline-fast unboxing pattern. We avoid calling
+/// the slower `js_nanbox_get_pointer` (which does the same thing in Rust)
+/// to keep concat hot-path overhead minimal.
+pub(crate) fn lower_string_concat(
+    ctx: &mut FnCtx<'_>,
+    left: &Expr,
+    right: &Expr,
+) -> Result<String> {
+    // #6951: same hazard as `lower_string_coerce_concat` — the left operand is
+    // a heap string in an SSA register across the right operand's evaluation.
+    let (l_box, r_box, guard) = lower_operand_pair_rooted(ctx, left, right)?;
+    let blk = ctx.block();
+    // SSO-aware fast path: pass operands as NaN-boxed f64s directly to
+    // `js_string_concat_sso`, which keeps SSO operands inline (no
+    // materialise-to-heap defeat) and returns the result NaN-boxed —
+    // SSO when the total fits 5 bytes, heap-pointer otherwise. Saves up
+    // to 3 heap allocations per concat on hot paths like ABC451D's
+    // recursive `before + after` (1.4M concats with 1-9 byte operands).
+    let result = blk.call(
+        DOUBLE,
+        "js_string_concat_box",
+        &[(DOUBLE, &l_box), (DOUBLE, &r_box)],
+    );
+    temp_root_release(ctx, guard);
+    Ok(result)
+}
+
+/// Cap the per-call part count for the n-way fold. Must match the
+/// runtime's `MAX_PARTS` in `js_string_concat_chain`. 32 covers every
+/// realistic CSV / log-line / template chain in user code.
+const CONCAT_CHAIN_MAX_PARTS: usize = 32;
+
+/// Try to flatten a left-spine of `Binary { Add }` nodes where every Add
+/// has at least one statically-string operand. Returns the parts in
+/// left-to-right (source-order) order. Returns `None` if the chain is
+/// shorter than the existing pairwise fast path's preference, has too
+/// many parts, or contains an Add node where neither side is statically
+/// string (which would risk numeric semantics under JS spec).
+///
+/// Caller passes the OUTERMOST Add's children. If the outermost Add's
+/// left child is itself a string-shaped Add, we recurse into it; right
+/// children are always leaves in our flat representation.
+pub(crate) fn flatten_string_add_chain<'a>(
+    ctx: &FnCtx<'_>,
+    left: &'a Expr,
+    right: &'a Expr,
+) -> Option<Vec<&'a Expr>> {
+    use perry_hir::BinaryOp;
+
+    let mut parts: Vec<&Expr> = Vec::with_capacity(8);
+    parts.push(right);
+
+    // Walk down the left spine. At each step, the current `cur` was the
+    // left child of an Add we already accepted — so we know `cur + ...`
+    // is string-shaped at the level above. We need each Add we descend
+    // INTO to itself be string-shaped (≥1 statically-string operand), so
+    // the entire chain has unambiguous string semantics.
+    let mut cur: &Expr = left;
+    loop {
+        match cur {
+            Expr::Binary {
+                op: BinaryOp::Add,
+                left: l,
+                right: r,
+            } => {
+                let l_str = crate::type_analysis::is_definitely_string_expr(ctx, l);
+                let r_str = crate::type_analysis::is_definitely_string_expr(ctx, r);
+                if !l_str && !r_str {
+                    // Stop the descent — this Add isn't unambiguously
+                    // string-shaped. Treat the entire `cur` subtree as
+                    // one opaque part.
+                    parts.push(cur);
+                    break;
+                }
+                parts.push(r);
+                cur = l;
+                if parts.len() >= CONCAT_CHAIN_MAX_PARTS {
+                    return None;
+                }
+            }
+            _ => {
+                parts.push(cur);
+                break;
+            }
+        }
+    }
+
+    parts.reverse();
+    Some(parts)
+}
+
+/// Lower a flat parts list to a single `js_string_concat_chain` call.
+/// Each part is lowered to its NaN-boxed value, then stored into a
+/// stack-allocated `[CONCAT_CHAIN_MAX_PARTS x double]` buffer; we pass
+/// the base pointer + N to the runtime helper, which produces a single
+/// allocation containing the entire concatenated result.
+///
+/// The buffer is fixed-size (always sized to MAX_PARTS) and hoisted to
+/// the function entry block via `alloca_entry_array`. A non-entry-block
+/// alloca lowers to a runtime `sub %rsp, N` with no matching restore;
+/// inside a loop body that's a stack leak (issue #167 — same shape that
+/// blew up `buf.readInt32BE` in tight loops). Function-entry allocas
+/// run once at prologue and the slot dominates every reachable use.
+/// One per-function buffer is shared across all chain call sites — fine
+/// because each chain call writes its parts and immediately calls into
+/// the runtime helper before any other call site can clobber the slots.
+pub(crate) fn lower_string_concat_chain(ctx: &mut FnCtx<'_>, parts: &[&Expr]) -> Result<String> {
+    debug_assert!(parts.len() >= 2);
+    debug_assert!(parts.len() <= CONCAT_CHAIN_MAX_PARTS);
+
+    // Lower each part first (in source order); side effects must fire
+    // left-to-right per JS spec. #6951: that ordering is exactly what makes
+    // every earlier part a heap value in an SSA register across every later
+    // part's evaluation — this is the template-literal / log-line shape, and
+    // one allocating interpolation was enough to sweep the parts already
+    // lowered. Parts that nothing allocating follows emit no rooting calls.
+    let (lowered, guard) = lower_exprs_rooted(ctx, parts)?;
+
+    let n = lowered.len();
+    // Hoist the buffer to the function entry block. Issue #167.
+    let buf_reg = ctx.func.alloca_entry_array(DOUBLE, CONCAT_CHAIN_MAX_PARTS);
+    let blk = ctx.block();
+    for (i, val) in lowered.iter().enumerate() {
+        let slot = blk.gep(DOUBLE, &buf_reg, &[(I64, &format!("{}", i))]);
+        blk.store(DOUBLE, val, &slot);
+    }
+    // Pass the array's base pointer as i64 (codegen ABI uses i64 for
+    // raw pointer args matching the existing `js_string_concat` shape).
+    let base_i64 = blk.next_reg();
+    blk.emit_raw(format!("{} = ptrtoint ptr {} to i64", base_i64, buf_reg));
+
+    let result_handle = blk.call(
+        I64,
+        "js_string_concat_chain",
+        &[(I64, &base_i64), (I32, &format!("{}", n))],
+    );
+    let boxed = nanbox_string_inline(blk, &result_handle);
+    temp_root_release(ctx, guard);
+    Ok(boxed)
+}
